+++ binaries/pico_limbo/src/server/mod.rs
@@ -6,3 +6,4 @@ pub mod game_mode;
 pub mod game_profile;
 mod named_packet;
 pub mod network;
+mod shutdown_signal;